requirements.txt
asgiref==3.7.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.7
django-environ==0.11.2
filelock==3.13.1
gunicorn==21.2.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
pyzbar==0.1.9
qrcode==7.4.2
reportlab==4.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.25.0
whitenoise==6.6.0
